Updated implemented assetStore
deleted all "ok1"

+++ src/ECS/ECS.cpp
@@ -36,7 +36,7 @@ Entity Registry::CreateEntity() {
     
     entitiesToBeAdded.insert(entity);
 
-    if (entityId >= entityComponentSignatures.size()) {
+    if (entityId >= static_cast<int>(entityComponentSignatures.size())) {
         entityComponentSignatures.resize(entityId + 1);
     }
