+++ src/ECS/ECS.h
@@ -234,7 +234,7 @@ void Registry::AddComponent(Entity entity, TArgs&& ...args){
     const auto componentId = Component<TComponent>::GetId();
     const auto entityId = entity.Getid();
 
-    if (componentId >= componentPools.size()){
+    if (componentId >= static_cast<int>(componentPools.size())){
         componentPools.resize(componentId + 1, nullptr);
     }
